+++ src/pages/index.js
@@ -1,69 +1,78 @@
 import React from "react"
-// import { Link } from "gatsby"
-import { graphql } from "gatsby"
-import AboutSection from "../components/AboutSection"
-import ExpertiseSection from "../components/ExpertiseSection"
-import EducationSection from "../components/EducationSection"
-import EmploymentSection from "../components/EmploymentSection"
-import SpeakingSection from "../components/SpeakingSection"
-import SkillsSection from "../components/SkillsSection"
-import OrganizationsSection from "../components/OrganizationsSection"
-import InterestsSection from "../components/InterestsSection"
-import OpensourceSection from "../components/OpensourceSection"
-import Layout from "../components/layout"
-// import Image from "../components/image"
-import SEO from "../components/seo"
+import { Link, graphql } from "gatsby"
 
-export const query = graphql`
-  query EmploymentQuery {
-    diceappImages: allFile(filter: { name: { regex: "/^diceapp/" } }) {
-      edges {
-        node {
-          relativePath
-          childImageSharp {
-            fluid(maxWidth: 800) {
-              ...GatsbyImageSharpFluid
-            }
-          }
-        }
+// import Bio from "../components/bio"
+import Layout from "../components/layout"
+import SEO from "../components/seo"
+import Section from "../components/Section"
+import AboutSection from "../components/AboutSection"
+// import { rhythm } from "../utils/typography"
+
+class BlogIndex extends React.Component {
+  render() {
+    const { data } = this.props
+    const siteTitle = data.site.siteMetadata.title
+    const posts = data.allMarkdownRemark.edges
+
+    return (
+      <Layout location={this.props.location} title={siteTitle}>
+        <SEO
+          title="All posts"
+          keywords={[`blog`, `gatsby`, `javascript`, `react`]}
+        />
+        <AboutSection />
+        <Section>
+        {posts.map(({ node }) => {
+          const title = node.frontmatter.title || node.fields.slug
+          return (
+            <div key={node.fields.slug}>
+              <h3
+                style={{
+                //   marginBottom: rhythm(1 / 4),
+                }}
+              >
+                <Link style={{ boxShadow: `none` }} to={node.fields.slug}>
+                  {title}
+                </Link>
+              </h3>
+              <small>{node.frontmatter.date}</small>
+              <p
+                dangerouslySetInnerHTML={{
+                  __html: node.frontmatter.description || node.excerpt,
+                }}
+              />
+            </div>
+          )
+        })}
+        </Section>
+      </Layout>
+    )
+  }
+}
+
+export default BlogIndex
+
+export const pageQuery = graphql`
+  query {
+    site {
+      siteMetadata {
+        title
       }
     }
-    atomicdexImages: allFile(filter: { name: { regex: "/^atomicdex/" } }) {
+    allMarkdownRemark(sort: { fields: [frontmatter___date], order: DESC }) {
       edges {
         node {
-          relativePath
-          childImageSharp {
-            fluid(maxWidth: 800) {
-              ...GatsbyImageSharpFluid
-            }
+          excerpt
+          fields {
+            slug
+          }
+          frontmatter {
+            date(formatString: "MMMM DD, YYYY")
+            title
+            description
           }
         }
       }
     }
   }
 `
-
-const IndexPage = ({ classes, data }) => (
-  <Layout>
-    <SEO
-      title="Home"
-      keywords={[`resume`, `nam`, `hoang`, `personal`, `website`]}
-    />
-    <AboutSection />
-    <ExpertiseSection />
-    <EmploymentSection data={data} />
-    <EducationSection />
-    <SkillsSection />
-    <OpensourceSection />
-    <SpeakingSection />
-    <OrganizationsSection />
-    <InterestsSection />
-    {/* <div style={{ maxWidth: `300px`, marginBottom: `1.45rem` }}>
-      <Image />
-    </div>
-    <Link to="/page-2/">Go to page 2</Link> */}
-  </Layout>
-)
-
-export default IndexPage
-// update skills, speaking, organizations, interests section
